requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
google-cloud-storage==1.29.0